ws_test_infra/infrastructure/requirements.txt
aws-cdk-lib>2.0,<3.0
aws-xray-sdk==2.12.0